# contracts/walksafe.openapi.json
{
  "components": {
    "schemas": {
...
      "AccountDeletionRequestV2": {
        "additionalProperties": false,
        "properties": {
          "client_revision": {
            "maximum": 9.223372036854776e+18,
            "minimum": 1.0,
            "title": "Client Revision",
            "type": "integer"
          },
          "confirmation": {
            "const": "DELETE_MY_ACCOUNT",
            "title": "Confirmation",
            "type": "string"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 16,
            "pattern": "^[A-Za-z0-9_-]{16,128}$",
            "title": "Request Id",
            "type": "string"
          },
          "schema_version": {
            "const": "walksafe.account-deletion-request.v2",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "request_id",
          "client_revision",
          "confirmation"
        ],
        "title": "AccountDeletionRequestV2",
        "type": "object"
      },
...
      "DeviceDeletionEvidenceV2": {
        "additionalProperties": false,
        "properties": {
          "client_revision": {
            "maximum": 9.223372036854776e+18,
            "minimum": 1.0,
            "title": "Client Revision",
            "type": "integer"
          },
          "completed_at": {
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$",
            "title": "Completed At",
            "type": "string"
          },
          "evidence_id": {
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{7,127}$",
            "title": "Evidence Id",
            "type": "string"
          },
          "evidence_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Evidence Sha256",
            "type": "string"
          },
          "expected_status_revision": {
            "maximum": 9.223372036854776e+18,
            "minimum": 1.0,
            "title": "Expected Status Revision",
            "type": "integer"
          },
          "installation_id": {
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{7,127}$",
            "title": "Installation Id",
            "type": "string"
          },
          "item": {
            "const": "device_untransmitted_data",
            "title": "Item",
            "type": "string"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 16,
            "pattern": "^[A-Za-z0-9_-]{16,128}$",
            "title": "Request Id",
            "type": "string"
          },
          "request_receipt_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Request Receipt Sha256",
            "type": "string"
          },
          "result": {
            "enum": [
              "DELETED",
              "NOT_FOUND",
              "FAILED"
            ],
            "title": "Result",
            "type": "string"
          },
          "schema_version": {
            "const": "walksafe.device-deletion-evidence.v2",
            "title": "Schema Version",
            "type": "string"
          },
          "tombstone_id": {
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{7,127}$",
            "title": "Tombstone Id",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "request_id",
          "tombstone_id",
          "request_receipt_sha256",
          "installation_id",
          "evidence_id",
          "client_revision",
          "expected_status_revision",
          "item",
          "result",
          "completed_at",
          "evidence_sha256"
        ],
        "title": "DeviceDeletionEvidenceV2",
        "type": "object"
      },
...
      "PrivacyErrorResponseV2": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/PrivacyErrorDetailV2"
          }
        },
        "required": [
          "detail"
        ],
        "title": "PrivacyErrorResponseV2",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "WalkSafeAccountGeneration": {
        "description": "Positive account generation bound by the gateway actor assertion.",
        "in": "header",
        "name": "X-WalkSafe-Account-Generation",
        "type": "apiKey"
      },
      "WalkSafeActorAssertion": {
        "description": "Short-lived gateway assertion binding the actor to its authenticated role.",
        "in": "header",
        "name": "X-WalkSafe-Actor-Assertion",
        "type": "apiKey"
      },
      "WalkSafeActorId": {
        "description": "Named actor for secured, rate-limited, or audited operations.",
        "in": "header",
        "name": "X-WalkSafe-Actor-Id",
        "type": "apiKey"
      },
      "WalkSafeAdminAppKind": {
        "description": "Required exact value: ADMIN_ANDROID.",
        "in": "header",
        "name": "X-WalkSafe-App-Kind",
        "type": "apiKey",
        "x-walksafe-required-value": "ADMIN_ANDROID"
      },
      "WalkSafeAdminAudience": {
        "description": "Required exact value: walksafe-admin-api.",
        "in": "header",
        "name": "X-WalkSafe-Audience",
        "type": "apiKey",
        "x-walksafe-required-value": "walksafe-admin-api"
      },
      "WalkSafeAdminBearer": {
        "bearerFormat": "opaque per-device session",
        "description": "Database-backed administrator session; static administrator tokens are rejected.",
        "scheme": "bearer",
        "type": "http"
      },
      "WalkSafeAdminDeviceChallengeId": {
        "description": "Canonical UUID of the fresh one-use device challenge.",
...
  },
  "info": {
    "title": "WalkSafe Assist API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/privacy/account-deletions": {
      "post": {
        "operationId": "create_account_deletion_privacy_account_deletions_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-walksafe-actor-id",
            "required": true,
            "schema": {
              "title": "X-Walksafe-Actor-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-walksafe-account-generation",
            "required": true,
            "schema": {
              "title": "X-Walksafe-Account-Generation",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-walksafe-deletion-access-pre-digest",
            "required": true,
            "schema": {
              "title": "X-Walksafe-Deletion-Access-Pre-Digest",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-walksafe-deletion-tombstone-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Walksafe-Deletion-Tombstone-Id"
            }
          },
          {
            "in": "header",
            "name": "x-walksafe-actor-assertion",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Walksafe-Actor-Assertion"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDeletionRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDeletionStatusV2"
                }
              }
            },
            "description": "Idempotent replay of an accepted deletion request."
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDeletionStatusV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "WalkSafeAccountGeneration": [],
            "WalkSafeActorAssertion": [],
            "WalkSafeActorId": [],
            "WalkSafeDeletionAccessPreDigest": [],
            "WalkSafeFieldToken": []
          }
        ],
        "summary": "Create Account Deletion",
        "tags": [
          "privacy"
        ],
        "x-walksafe-actor-identity-required": true,
        "x-walksafe-deletion-capability-scope": "account-deletion-only",
        "x-walksafe-required-role": "field"
      }
    },
    "/privacy/account-deletions/{request_id}/device-evidence": {
      "post": {
        "operationId": "create_device_deletion_evidence_privacy_account_deletions__request_id__device_evidence_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9_-]{16,128}$",
              "title": "Request Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-walksafe-actor-id",
            "required": true,
            "schema": {
              "title": "X-Walksafe-Actor-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-walksafe-account-generation",
            "required": true,
            "schema": {
              "title": "X-Walksafe-Account-Generation",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-walksafe-deletion-access-pre-digest",
            "required": true,
            "schema": {
              "title": "X-Walksafe-Deletion-Access-Pre-Digest",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-walksafe-deletion-tombstone-id",
            "required": true,
            "schema": {
              "title": "X-Walksafe-Deletion-Tombstone-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-walksafe-actor-assertion",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Walksafe-Actor-Assertion"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceDeletionEvidenceV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDeletionStatusV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "WalkSafeAccountGeneration": [],
            "WalkSafeActorAssertion": [],
            "WalkSafeActorId": [],
            "WalkSafeDeletionAccessPreDigest": [],
            "WalkSafeDeletionTombstoneId": [],
            "WalkSafeFieldToken": []
          }
        ],
        "summary": "Create Device Deletion Evidence",
        "tags": [
          "privacy"
        ],
        "x-walksafe-actor-identity-required": true,
        "x-walksafe-deletion-capability-scope": "account-deletion-only",
        "x-walksafe-required-role": "field"
      }
    },
    "/privacy/account-deletions/{request_id}/status": {
      "get": {
        "operationId": "account_deletion_status_privacy_account_deletions__request_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9_-]{16,128}$",
              "title": "Request Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-walksafe-actor-id",
            "required": true,
            "schema": {
              "title": "X-Walksafe-Actor-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-walksafe-account-generation",
            "required": true,
            "schema": {
              "title": "X-Walksafe-Account-Generation",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-walksafe-deletion-access-pre-digest",
            "required": true,
            "schema": {
              "title": "X-Walksafe-Deletion-Access-Pre-Digest",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-walksafe-deletion-tombstone-id",
            "required": true,
            "schema": {
              "title": "X-Walksafe-Deletion-Tombstone-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-walksafe-actor-assertion",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Walksafe-Actor-Assertion"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDeletionStatusV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyErrorResponseV2"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "WalkSafeAccountGeneration": [],
            "WalkSafeActorAssertion": [],
            "WalkSafeActorId": [],
            "WalkSafeDeletionAccessPreDigest": [],
            "WalkSafeDeletionTombstoneId": [],
            "WalkSafeFieldToken": []
          }
        ],
        "summary": "Account Deletion Status",
        "tags": [
          "privacy"
        ],
        "x-walksafe-actor-identity-required": true,
        "x-walksafe-deletion-capability-scope": "account-deletion-only",
        "x-walksafe-required-role": "field"
      }
    },
...
  },
...
}